app/pyproject.toml
[project]
name = "app"
version = "1.0.0"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
  "fastapi[standard]>=0.115.12",
  "httpx>=0.28.1",
  "opentelemetry-api>=1.31.1",
  "opentelemetry-exporter-otlp>=1.31.1",
  "opentelemetry-instrumentation-fastapi>=0.52b1",
  "opentelemetry-instrumentation-httpx>=0.52b1",
  "opentelemetry-instrumentation-logging>=0.52b1",
  "opentelemetry-sdk>=1.31.1",
]
